Develop

+++ cllc-public-app/ViewModels/InventorySalesReport.cs
@@ -17,15 +17,15 @@ namespace Gov.Lclb.Cllb.Public.ViewModels
         public int? lostReductions { get; set; }
         public int? otherReductions { get; set; }
         public int? closingNumber { get; set; }
-        public double closingValue { get; set; }
-        public double closingWeight { get; set; }
+        public decimal closingValue { get; set; }
+        public decimal closingWeight { get; set; }
         // only used for seeds
         public int? totalSeeds { get; set; }
 
         // Sales
         public int? totalSalesToConsumerQty { get; set; }
-        public double totalSalesToConsumerValue { get; set; }
+        public decimal totalSalesToConsumerValue { get; set; }
         public int? totalSalesToRetailerQty { get; set; }
-        public double totalSalesToRetailerValue { get; set; }
+        public decimal totalSalesToRetailerValue { get; set; }
     }
 }
